# shared/contracts/openapi/rpml-web-app.v1.yaml
openapi: 3.1.0
info:
  title: RPML Backend
  description: 'RPML Web App HTTP API (contract snapshot: shared/contracts/openapi/rpml-web-app.v1.yaml).'
  version: 1.0.0
paths:
...
  /api/v1/demo/seed:
    post:
      tags:
      - demo
      summary: Заполнить демо-данные
      description: Идемпотентно создает/обновляет демо-сценарий и связанные данные
        для текущего пользователя, чтобы можно было сразу запускать оптимизацию.
      operationId: post_demo_seed_api_v1_demo_seed_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DemoSeedRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoSeedResponse'
        '401':
          description: Пользователь не аутентифицирован.
        '422':
          description: Ошибка валидации демо-данных.
      security:
      - HTTPBearer: []
  /api/v1/demo/instances:
    get:
      tags:
      - demo
      summary: Список доступных demo-инстансов
      description: Возвращает каталог bundled demo-сценариев из `core/rpml/result_samples`,
        которые можно загрузить через `/api/v1/demo/seed`.
      operationId: get_demo_instances_api_v1_demo_instances_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DemoInstanceSummary'
                type: array
                title: Response Get Demo Instances Api V1 Demo Instances Get
        '401':
          description: Пользователь не аутентифицирован.
      security:
      - HTTPBearer: []
  /api/v1/demo/runs/{run_id}/top-savings:
    get:
      tags:
      - demo
      summary: Топ инстансов по экономии MILP
      description: Читает checkpoint эксперимента run и возвращает top-N инстансов
        по процентной экономии MILP относительно Avalanche или Snowball.
      operationId: get_demo_run_top_savings_api_v1_demo_runs__run_id__top_savings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: metric
        in: query
        required: false
        schema:
          enum:
          - avalanche
          - snowball
          type: string
          description: Базовая стратегия для сравнения.
          default: avalanche
          title: Metric
        description: Базовая стратегия для сравнения.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Количество строк в топе.
          default: 10
          title: Limit
        description: Количество строк в топе.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoRunTopSavingsResponse'
        '401':
          description: Пользователь не аутентифицирован.
        '404':
          description: Run/checkpoint не найден.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DemoInstanceSummary:
      properties:
        code:
          type: string
          title: Code
          description: Код сценария, который сохраняется в `scenario_profiles.code`.
        source_file:
          type: string
          title: Source File
          description: Имя bundled JSON-файла в `core/rpml/result_samples`.
        horizon_months:
          type: integer
          title: Horizon Months
          description: Горизонт сценария в месяцах.
        loans_count:
          type: integer
          title: Loans Count
          description: Количество долгов в bundled сценарии.
        has_dat_file:
          type: boolean
          title: Has Dat File
          description: Есть ли рядом исходный `.dat` с реальными ставками.
        available_baselines:
          items:
            type: string
          type: array
          title: Available Baselines
          description: Baseline-стратегии, присутствующие в bundled summary.
      type: object
      required:
      - code
      - source_file
      - horizon_months
      - loans_count
      - has_dat_file
      title: DemoInstanceSummary
    DemoRunTopSavingsItem:
      properties:
        instance_name:
          type: string
          title: Instance Name
          description: Имя инстанса из checkpoint.
        savings_pct:
          type: number
          title: Savings Pct
          description: Экономия MILP vs выбранная baseline-стратегия в процентах.
        savings_abs:
          type: number
          title: Savings Abs
          description: Абсолютная экономия MILP vs выбранная baseline-стратегия.
        optimal_cost:
          type: number
          title: Optimal Cost
          description: Стоимость MILP.
        baseline_cost:
          type: number
          title: Baseline Cost
          description: Стоимость baseline-стратегии.
        optimal_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Optimal Status
          description: Статус MILP решения для инстанса.
      type: object
      required:
      - instance_name
      - savings_pct
      - savings_abs
      - optimal_cost
      - baseline_cost
      title: DemoRunTopSavingsItem
    DemoRunTopSavingsResponse:
      properties:
        run_id:
          type: string
          title: Run Id
        metric:
          type: string
          enum:
          - avalanche
          - snowball
          title: Metric
        checkpoint_path:
          type: string
          title: Checkpoint Path
        total_instances:
          type: integer
          title: Total Instances
        items:
          items:
            $ref: '#/components/schemas/DemoRunTopSavingsItem'
          type: array
          title: Items
      type: object
      required:
      - run_id
      - metric
      - checkpoint_path
      - total_instances
      - items
      title: DemoRunTopSavingsResponse
    DemoSeedRequest:
      properties:
        scenario_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Scenario Code
          description: Код bundled demo-сценария из `/api/v1/demo/instances`. Если
            не указан, используется стандартный demo-инстанс.
      type: object
      title: DemoSeedRequest
    DemoSeedResponse:
      properties:
        ok:
          type: boolean
          title: Ok
        scenario_code:
          type: string
          title: Scenario Code
        debts_count:
          type: integer
          title: Debts Count
      type: object
      required:
      - ok
      - scenario_code
      - debts_count
      title: DemoSeedResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...